+++ src/test/converter2/issues/gh2937/index.ts
@@ -1,0 +1,2 @@
+export const excluded = true;
+export { notExcluded } from "./not-excluded.js";